+++ spi_nand_flash/src/nand_flash_blockdev.c
@@ -291,8 +291,6 @@ static esp_err_t nand_flash_blockdev_release(esp_blockdev_handle_t handle)
 #ifdef CONFIG_IDF_TARGET_LINUX
     res = nand_emul_deinit(dev_handle);
 #endif
-    free(dev_handle->work_buffer);
-    free(dev_handle->read_buffer);
     free(dev_handle->temp_buffer);
     if (dev_handle->mutex) {
         vSemaphoreDelete(dev_handle->mutex);
@@ -325,8 +323,6 @@ esp_err_t nand_flash_get_blockdev(spi_nand_flash_config_t *config, esp_blockdev_
 
     esp_blockdev_t *blockdev = (esp_blockdev_t *) heap_caps_calloc(1, sizeof(esp_blockdev_t), MALLOC_CAP_DEFAULT);
     if (blockdev == NULL) {
-        free(handle->work_buffer);
-        free(handle->read_buffer);
         free(handle->temp_buffer);
         if (handle->mutex) {
             vSemaphoreDelete(handle->mutex);